Moved pug-resolve-plugin to shared/
