+++ shared/pug.js
@@ -3,7 +3,7 @@ var pug = require('pug');
 var _ = require('lodash');
 
 module.exports = function(config) {
-  var pugResolvePlugin = require('../webapplication/lib/pug-resolve-plugin')(config);
+  var pugResolvePlugin = require('./pug-resolve-plugin')(config);
   var viewsPath = path.join(__dirname, 'views', '*');
 
   // Resolve non-absolute paths
